+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning].
   of configuration.
   Showing the list screen only in development presents opportunities for bugs to crop up.
   But with different behaviors, they are hard to spot in development.
+- Config edit screen now shows the config key on the tooltip for each entry.
 
 ### Deprecated